Updating driveraker progress for today
The if-then statements which find and save variables should be saved in a more modular way. Check the code for notes.

I need to write a for/while loop to detect all in-line headers, images, and image captions. Check the code for notes.

This is my progress for #2 today. I should finish this sometime soon